refactor: add real-time audio transcription node and SRT generation capabilities
