+++ src/audio_transcription/__init__.py
@@ -1,0 +1,1 @@
+"""Audio transcription package for ComfyUI real-time transcription capabilities."""
